+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.3.3] - 2020-04-30
 
 ### Added
 
@@ -79,6 +79,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.2.2]: https://github.com/dogmatiq/aperture/releases/tag/v0.2.2
 [0.3.0]: https://github.com/dogmatiq/aperture/releases/tag/v0.3.0
 [0.3.1]: https://github.com/dogmatiq/aperture/releases/tag/v0.3.1
+[0.3.2]: https://github.com/dogmatiq/aperture/releases/tag/v0.3.2
+[0.3.3]: https://github.com/dogmatiq/aperture/releases/tag/v0.3.3
 
 [Unreleased]: https://github.com/dogmatiq/aperture